requirements.txt
flask
rembg
pillow
gunicorn
